Update `Set all tags` to include unique tag ref
The line we've added does a check to ensure this is a push to `main` (we only add a unique docker tag for these and not when a new git tag is pushed). If it is it generates a new tag based on the GHCR repo address, the image name and the output of `git describe --always --tags`.

By adding it into the existing array of tags the `build-push-action` will do the rest!