Create `DeprecatedRoutePlugin` (#544)
As part of our versioning strategy, we wish to be able to mark routes as deprecated and optionally provide the route that will be succeeding them. We do this with DeprecatedRoutePlugin, which we implement here.

A route can be marked as deprecated by adding a deprecation object to its route definition:

{
  method: 'GET',
  path: '/deprecated',
  handler: Controller.handler,
  options: {
    app: {
      deprecation: {
      }
    }
  }
}
It is entirely valid that this object can be empty; its presence is enough to mark the route as deprecated. A deprecated route will have the header deprecation: true added to the response.

Optionally, the route replacing it (its "successor") can be specified:

{
  method: 'GET',
  path: '/deprecated',
  handler: Controller.handler,
  options: {
    app: {
      deprecation: {
        successor: '/not-deprecated'
      }
    }
  }
}
This would add the header link: </not-deprecated>; rel="successor-version" to the link response, alongside deprecation: true.

Note that the commit history shows that we initially created this as DeprecatedEndpointPlugin, and implement the logic in a separate DeprecatedEndpointService. During development we decided the logic could live in the plugin, and that DeprecatedRoutePlugin was a more suitable name.


* Initial registration of plugin

* Create `DeprecatedEndpointService`

* Amend `addPublicRoute` helper

We need the ability to add options to routes in order to test the plugin, so we update `RouteHelper.addPublicRoute()` accordingly

* Change plugin registration order

If we register the plugin too early then we run into problems with adding the headers to error responses, etc. Registering the plugin later avoids these.

* Develop `DeprecatedEndpointPlugin`

* Document plugin and service

* Move functionality into plugin

We decided that the functionality was light enough that it didn't require a separate service. We therefore Remove `DeprecatedEndpointService` and move its functionality into `DeprecatedEndpointPlugin`

* Rename from `Endpoint` to `Route`

After consideration we realised that we weren't strictly being correct when we referred to endpoints as being deprecated; we are deprecating individual _routes_ instead. We therefore rename the plugin from `DeprecatedEndpointPlugin` to `DeprecatedRoutePlugin`.

* Fix SonarCloud errors

SonarCloud reminded us that since our arrow functions in the plugin are only a single statement, we don't need curly braces or `return`, so we remove them.