Fix uppercase org name in unique docker tag
In [Generate and add unique tag in docker workflow](https://github.com/DEFRA/sroc-charging-module-api/pull/543) we overlooked that the `GHC_REGISTRY` contains the org name 'DEFRA'. Docker tags must be all lowercase and the `metadata-action` takes care of that. Our custom command isn't though. So, rather than make something that is never going to change even more complicated we just replace some of the code with the repo name we need.