Rename show endpoints to view (#547)
https://github.com/DEFRA/sroc-service-team/issues/68

When we first started building the project we focused on the admin endpoints, specifically `regimes` and `authorised_systems`. They were relatively simple and we needed the underlying data anyway in order to build the rest of the service.

But we were coming from a Ruby on Rails background which meant we used their conventions. Specifically, `show` for responding to a `GET` request.

In later endpoints, we switched to `view` as this was the terminology used by the rest of the team and in our documentation. So, we have an inconsistency in endpoint naming that this change resolves.