Set up customer files endpoint and dummy service (#549)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-57

The first stage of the customer file functionality is to set up the endpoint, route, and a dummy service.

We immediately hit an issue; we want to call this service ListCustomerFilesService, but we already have a service with this name, which we use for the test endpoint. Rather than come up with an alternative name (eg. FetchCustomerFilesService) we rename the existing service to TestListCustomerFilesService which makes its purpose clear. We then proceed to create our dummy service, which simply returns true for now.

Then we create our CustomerFilesController and route GET /v2/{regimeId}/customer-files. The route is named to match the admin routes, and the controller named to match the route; both these names are up for discussion however.


* Rename `ListCustomerFilesService`

The existing `ListCustomerFilesService` is for test purposes only, and its name clashes with the intended name for our new service. Rather than name the new service something along the lines of `FetchCustomerFilesService`, we rename the existing service to `TestListCustomerFilesService` make its purpose clear.

* Create `ListCustomerFilesService`

* Create `CustomerFilesController`

* Create `GET /v2/{regimeId}/customer-files` route

* Put unit tests in place

* Fix unit test issue