Implement `ListCustomerFilesService` (#554)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907/backlog?selectedIssue=CMEA-166

We continue with this feature by implementing the ListCustomerFilesService, along with the ListCustomerFilesPresenter it uses to format its output. Note that we do not yet enforce the acceptance criteria that the output should default to 30 days; this will be handled in a future branch.

As part of development, we have amended CustomerHelper.addCustomerFile() to accept an export date, defaulting to the current date to avoid breaking existing tests. While doing this, we noticed that it was not setting the customer file status, so we fixed this.

Finally, we remove the service from CustomerFilesController -- after developing the service fully, the tests now fail as they don't perform the required setup. Rather than continuing to develop the controller in this branch, we remove the service with the intent of reinstating it in a future branch.

* Add export date to `CustomerHelper`

The customer helper defaults the export date to `Date.now()` when creating an `exported` customer file. For testing purposes we need to create files in the past, so we amend the helper to optionally accept a value for the export date.

* `CustomerHelper` now sets status

While updating the customer helper we realised that it wasn't setting the status of the created customer file, so we fix this here.

* Add `daysAgoDate` to `GeneralHelper`

We add a new helper to give us the date a given number of days ago, which we will use to create specific customer file records

* Initial implementation

* `ListCustomerFilesService` returns customer refs

We need to provide the customer references in each exported file, so we use `withGraphFetched()` to bring in data from the `exportedCustomers` table.

* Implement `ListCustomerFilesPresenter`

This presenter formats the output of `ListCustomerFilesService`

* Update service to use new presenter

* Update service docs

* Additional test case

We add an additional test case just to confirm that files exported on the specified day are returned. ie. if we specify we want files exported 1 day ago, we will get the file exported yesterday.

* Remove service from controller

The controller test fails now that we have fully developed the service. Rather than continuing to develop the controller in this branch, we temporarily remove the service from the controller.