Amend version in `.nvmrc` to from `14.6.0` to `14`
We use `alpine-14` as our base Docker image, which will automatically build it on the latest version of Node 14. We are therefore updating `.nvmrc` to also specify version 14 (rather than a specific version, ie. 14.6.0) to ensure that GitHub CI runs tests against the same version of Node that the Docker image has.