Expand info returned in customer-files endpoint (#558)
https://eaflood.atlassian.net/browse/CMEA-57

In [Implement ListCustomerFilesService](https://github.com/DEFRA/sroc-charging-module-api/pull/554) we built the service which will be used by the new public version of our customer files endpoint.

The intent is now that we have a public version of the endpoint we should be able to delete `/admin/test` version. However, we've spotted that the 2 return very different data. The `/admin/test` one returns

```json
[
    {
        "id": "13273263-787c-4868-9fd4-1a6b8a25c415",
        "regimeId": "48faa47b-e1fc-4892-ab8c-5d6862f1a99f",
        "region": "A",
        "fileReference": "nalac50001",
        "status": "exported",
        "exportedAt": "2021-09-21T22:40:45.398Z",
        "createdAt": "2021-09-21T22:40:44.859Z",
        "updatedAt": "2021-09-21T22:40:45.403Z"
    }
]
```

The new public `/v2/{regime}/` one returns only

```json
[
    {
        "fileReference": "nalac50001",
        "exportedCustomers": [
            "BB02BEEB"
        ]
    }
]
```

Ideally, before we delete the old one we should try and include some of that additional data in the public one.

So, this change updates the endpoint so it returns

```json
[
    {
        "id": "13273263-787c-4868-9fd4-1a6b8a25c415",
        "fileReference": "nalac50001",
        "status": "exported",
        "exportedAt": "2021-09-21T22:40:45.398Z",
        "exportedCustomers": [
            "BB02BEEB"
        ]
    }
]
```