Fix customer file head invalid date format (#557)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-166

Following a recent version release, a customer file was rejected by SSCL due to the header date being in the incorrect format -- the file header had the date in the format 19-SEPT-2021 when it should be 19-SEP-2021 (ie. it had 4 chars for the month when it should have been 3). Much head-scratching followed -- not only were we unable to reliably replicate this locally, but we hadn't even touched any of the code that formats this date since it was originally written, giving no reason why customer files produced at the start of September worked but those produced after the release didn't!

To cut a long story short -- we use toLocaleString() with the locale en-gb to give us the date in a format close to the one required (eg. 19 Sep 2021). However it seems that Node's implementation of this was bugged -- it was returning Sep as the month but it should have actually have been returning Sept all along! This bug was fixed in Node at some point between the recent release and the one before it, hence the reason the new version seemed to break things.

We have resolved this by using the locale en-nz in place of en-gb, as this consistently returns a 3-character month. And just to guard against this happening in future, we expand the unit test to cover every month.

* Change date formatting locale

* Amend version in `.nvmrc` to from `14.6.0` to `14`

We use `alpine-14` as our base Docker image, which will automatically build it on the latest version of Node 14. We are therefore updating `.nvmrc` to also specify version 14 (rather than a specific version, ie. 14.6.0) to ensure that GitHub CI runs tests against the same version of Node that the Docker image has.