Check bill run status before restoring it
When we delete a licence, we also check if the invoice it belongs to is empty. If it is then we call `DeleteInvoiceService`, which in turn checks if the bill run it belongs to is empty. If that's the case then it sets the bill run status to `initialised`. However this would then be overwritten when we restore the original bill run status. We therefore add a check to see if the bill run status was changed from `pending` before we restore the status.