Change based on review
When we changed from passing in the bill run id to just passing in the bill run, we missed the opportunity to simply take the status from the bill run instead of retrieving it from the db and then using its status. This was picked up in review and we correct it here.