Small refactor
Rather than retrieving the refreshed bill run status and checking if it is still `pending` to decide if we revert the status, we can simply attempt to revert the status anyway but use `.select('pending')` to determine if this takes place. Since this means there's no longer any point in a generic 'set bill run status' method, we can make it a more specific 'set bill run status to pending' method.