Fix deleting a deminimis invoice recalculation
https://eaflood.atlassian.net/browse/CMEA-171

After a discrepency with a bill run's summary was reported we tracked it down to an error when a deminimis invoice is deleted.

Deminimis invoices (debits £5 or less) are not included in the bill run summary. This means when it's generated they are not included in the invoice count and total value. So, when deleted they should not impact those values in the bill run. The problem is that we are updating the bill run when a deminimis invoice is deleted.

This change fixes the issue and ensures the bill run summary is correctly updated when a deminimis invoice is deleted.

**How to reproduce**

Create a bill run and add 5 invoices; 4 standard and 1 deminimis. Generate the bill run and it will show 4 invoices in the summary. The invoice value will be the total of the 4 non-deminimis invoices.

All good so far.

Then delete the deminimis invoice. The debit line count and values should change but the invoice values should not. The deminimis invoice was not included in the original calculation so it should not be affecting them.