Delete Licence not updating bill run status (#559)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-141

Unlike other endpoints which set off a background task and return a response immediately, Delete Licence does not update the bill run status to reflect the fact that an operation is running which could affect the bill run values. We therefore update DeleteLicenceService to do this.

One case to consider is when deleting the licence causes the invoice to be empty and therefore deleted, which in turn causes the bill run to be empty, in which case the status will be changed to initialised. We ensure that this is not overwritten by reverting to the original status by checking whether the status changed during deletion before restoring it.

Finally, we update the service's unit tests to use our new test helpers.


* Delete Licence not updating bill run status

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-141

* Refactor into `_deleteLicence()` method

Following the model of `DeleteBillRunService`, we want to move our deletion code into a separate method to keep the `try/catch` block nice and clean for when we start updating the status.

* Add `billRun` to parameters

We need to determine the current bill run status so we add `billRun` to the service's parameters

* Set and restore bill run status

* Check bill run status before restoring it

When we delete a licence, we also check if the invoice it belongs to is empty. If it is then we call `DeleteInvoiceService`, which in turn checks if the bill run it belongs to is empty. If that's the case then it sets the bill run status to `initialised`. However this would then be overwritten when we restore the original bill run status. We therefore add a check to see if the bill run status was changed from `pending` before we restore the status.

* Update tests to use new helpers

We take the opportunity to update the unit tests to use our new streamlined test helpers.

* Remove stray `.only`

* Fix comment

* Change based on review

When we changed from passing in the bill run id to just passing in the bill run, we missed the opportunity to simply take the status from the bill run instead of retrieving it from the db and then using its status. This was picked up in review and we correct it here.

* Small refactor

Rather than retrieving the refreshed bill run status and checking if it is still `pending` to decide if we revert the status, we can simply attempt to revert the status anyway but use `.select('pending')` to determine if this takes place. Since this means there's no longer any point in a generic 'set bill run status' method, we can make it a more specific 'set bill run status to pending' method.