Fix `ListCustomerFilesService` time boundary defect (#562)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-57

It was spotted during testing that any customer files exported at exactly midnight (ie. with a timestamp of 00:00:00.000 are not included in the service output. Although this is the edgiest of edge cases, it is still worth resolving, and it has a very simple cause and solution -- we were selecting the files to be included with .where('exportedAt', '>', startDate) when we should have been using .where('exportedAt', '>=', startDate).

* Amend test helper

We update the `daysAgoDate` test helper to optionally accept a time which we will use to create a file exported at exactly midnight, ie. 00:00:00.000

* Introduce failing test

We create a test that should pass (ie. to check that a file exported at exactly midnight is included), but doesn't.

* Fix boundary defect in `ListCustomerFilesService`

The defect has a simple cause; we select all files exported after 00:00:00.000, and not all files exported _on or after_ that time. We simply need to change a `>` to `>=` to resolve this.