Fix deleting invoice from initialised bill run (#568)
https://eaflood.atlassian.net/browse/CMEA-171

This all stems from finding out that when a deminimis invoice is deleted we were updating the bill run summary details when we shouldn't be. We then found the same would happen for zero value invoices

- [Fix deleting a deminimis invoice recalculation](https://github.com/DEFRA/sroc-charging-module-api/pull/561)
- [Fix deleting a zero value invoice recalculation](https://github.com/DEFRA/sroc-charging-module-api/pull/567)

Whilst checking those we hit on another realisation. The bill run summary (we refer to the credit note/invoice fields in a bill run as the bill run summary) is only set when a bill run is generated. But the `DeleteInvoiceService` is always patching these fields, deminimis and zero value being the exceptions.

We should not be attempting to update these fields if the bill run is not `initialised` as this could cause it to go into negative figures.

N.B. It's not supercritical. When the bill run is generated the values would be fixed. But it would be confusing prior to the bill run being generated which is why we deal with it now.