Update test bill run generator with trans. count
https://eaflood.atlassian.net/browse/CMEA-141

We've been trying to create a test where we can see the bill run status go to `pending` when a licence is deleted. Problem is, with a whole bunch of transactions the service is proving to fast to see it in our acceptance tests.

We have the bill run generator test endpoint. The only inssue is the current counts relate to number of invoices. The number of transactions per invoice is set in the code.

To give us a bit more flexibility, this change allows us to tell the generator how many transactions to add per invoice.

**Note**

It's not always a direct transalation. For example, if we set `transactionMultiplier` to 2 for a mixed invoice, we'll get 4 standard transactions and 2 credit transactions. This is because we are multiplying whatever we were adding before by the transaction multiplier.