Update test bill run generator with trans. multiplier (#570)
https://eaflood.atlassian.net/browse/CMEA-141

We've been trying to create a test where we can see the bill run status go to `pending` when a licence is deleted. Problem is, with a whole bunch of transactions the service is proving too fast to see it in our acceptance tests.

We have the bill run generator test endpoint. The only issue is the current counts relate to the number of invoices. The number of transactions per invoice is set in the code.

To give us a bit more flexibility, this change allows us to affect the number of transactions added per invoice by the generator.

It also adds a new invoice type of `standard` which just adds debit transactions.

```json
{
    "region": "A",
    "transactionMultiplier": 5,
    "mix": [
        { "type": "mixed-invoice", "count": 1 },
        { "type": "mixed-credit", "count": 1 },
        { "type": "zero-value", "count": 1 },
        { "type": "deminimis", "count": 1 },
        { "type": "minimum-charge", "count": 1 },
        { "type": "standard", "count": 1 }
    ]
}
```

**Note**

It's not always a direct translation. For example, if we set `transactionMultiplier` to 2 for a mixed invoice, we'll get 4 standard transactions and 2 credit transactions. This is because we are multiplying whatever we were adding before by the transaction multiplier.