Fix postcode should not be a required property
https://eaflood.atlassian.net/browse/CMEA-181

Currently, `postcode` is required in any customer change requests to the API. We think this decision was rooted in the fact it's a required field in the customer file that goes to SSL with the customer changes in it.

But the original A/C on the story for developing it for CM v2 was

> AC6. Where the postcode is missing from the incoming request, detail field 11 in the customer file is populated with a "." (full stop).

Having double checked the code in CM v1 we can see this is in the request validation

```javascript
postcode: stringValidator.max(60).allow('', null)
``

When it comes to the customer file the CM is expected to replace blank postcodes wit `.`. Again, we can see this is CM v1

```javascript
handlePostcodeValue (value) {
  if (value && value.trim().length) {
    return value
  } else {
    return '.'
  }
}
```

We have properly dropped the ⚽ with this one. Hence, this change to sort it out.