Fix postcode should not be a required property (#574)
https://eaflood.atlassian.net/browse/CMEA-181

Currently, `postcode` is required in any customer change requests to the API. We think this decision was rooted in the fact it's a required field in the customer file that goes to SSL with the customer changes in it.

But the original A/C on the story for developing it for CM v2 was

> AC6. Where the postcode is missing from the incoming request, detail field 11 in the customer file is populated with a "." (full stop).

Having double-checked the code in CM v1 we can see this is in the request validation

```javascript
postcode: stringValidator.max(60).allow('', null)
```

When it comes to the customer file the CM is expected to replace blank postcodes with `.`. Again, we can see this is CM v1

```javascript
handlePostcodeValue (value) {
  if (value && value.trim().length) {
    return value
  } else {
    return '.'
  }
}
```

We have properly dropped the ⚽  with this one. Hence, this change to sort it out.

** Notes

As we need to allow postcode to be empty on the request it makes sense to transfer our logic for defaulting the value to `.` there. We already use the same translator to default other customer change fields and our convention is to use translators for defaulting missing incoming values.

As the translator will default empty postcode fields to `.` there is no need to 'handle' empty postcodes in the customer file body presenter.

Doing it in the translator

- follows the convention we've set with other default values
- means we store in the DB what we actually send to SSCL
- removes the need for the extra logic in the customer file body presenter