Amend test helpers to refresh entity
The helpers act on and update the entity they are passed. This means that the entity must be freshed before use. For example, if the transaction helper is passed a licence to create a transaction on, and that licence is not refreshed before being passed to the transaction helper again, the counts will not reflect the first transaction. Therefore, we update the helpers to refresh the entity they are passed -- this prevents us forgetting to refresh the entity, which may cause issues with tests. If this slows down the tests then we will consider an additional 'refresh entity' flag (defaulting to `true`) which can be used when we know the entity doesn't need to be refreshed.