Fix Delete Licence not correctly updating bill run summary (#573)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-178

When deleting a licence, the bill run summary is not correctly updated in scenarios where the invoice changes "type". For example, from a debit/credit to zero value; from a credit to a debit; etc. In these cases, the bill run level values creditNoteCount, creditNoteValue, invoiceCount and invoiceValue were not correctly being adjusted to account for the change in type.

This was initially discovered in the following scenario:
- A debit transaction is created on licence L001;
- A zero-value transaction is created on licence L002;
- Licence L001 is deleted;
- The invoice count should be 0 but it remains 1.

(Note that the original defect report noted that creditNoteValue is also incorrect if the above steps are followed using a credit transaction; this turned out to be due to a typo in the code, which we quickly spotted and resolved.)

To fix the defect, we remove the invoice's previous net value (ie. the one it had prior to the licence being deleted) from the appropriate bill run credit/invoice value (determined by whether the invoice was previously an overall credit or debit) and decrement the appropriate bill run credit/invoice count by 1. We then add the invoice's new net value to the appropriate bill run value (determined by whether the invoice is now an overall credit or debit) and increment the appropriate count by 1. This accounts for all scenarios where the type changes, as well as those where it doesn't (as we are simply removing one value from the credit/debit field, and adding the new value back to it).

While doing this, we encountered a problem with the new test helpers; because they adjust the values of the entity they are passed, they must be given an updated/refreshed entity or else the adjusted values will be incorrect. For example, if a licence has a transaction with value 1000 added to it and then has another transaction of value 500 added to it without it being refreshed, the helper will add the 500 value to the invoice's initial value of 0. This can be avoided by refreshing the entity we pass to the helper prior to calling it; however it would be easy to forget to do this so we updated the helpers to automatically do it for us. This may slow the tests down a little; if it becomes too much to bear then we can look into this further (for example, adding an optional field, defaulting to true, so we can request that the entity isn't refreshed if we know this isn't necessary).

One thing to note is that refreshing an entity before use also applies to the services we call; this is why the tests have been amended to refresh the bill run after generation, and refresh each licence we pass to DeleteLicenceService.