Deminimis fix
The first defect found in testing was that the bill run summary was not correctly updated if the invoice the licence was deleted from was subject to deminimis. This was due to us removing the count and value from the bill run when these would start at 0 in the case of deminimis. We therefore check that the invoice was deminimis before updating. We also check whether the invoice becomes deminimis after deleting the licence, and if this is the case we don't add the count and value back to the summary.