Deleted invoice fix
The second defect found in testing was that if an invoice is deleted due to its only licence being deleted, the bill run summary was not being updated. We traced this to `DeleteInvoiceService`, which checks the bill run status to see if it's generated, and only updates the summary if it has been. However, due to our recent changes, calling this service from within `DeleteLicenceService` means that the bill run's status is `pending` at the point that it checks the status! We therefore update `DeleteInvoiceService` to accept an optional bill run status (which defaults to the current status) and pass in the bill run's initial status so it can correctly check whether the bill run is generated.