Delete Licence additional summary fixes (#581)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-178

Two main defects were found during testing of our revised approached to updating the bill run summary when a licence is deleted:

Defect 1 - deminimis

The bill run summary was not correctly updated if the invoice the licence was deleted from was subject to deminimis. This was due to us subtracting the count and value of the invoice from the bill run summary, when these would not have been included in the summary in the case of deminimis. We therefore introduce a check that the invoice was deminimis before updating. We also check whether the invoice becomes deminimis after deleting the licence, and if this is the case we don't add the count and value back to the summary.

Defect 2 - empty invoice deletion

The second defect was that if an invoice is deleted due to its only licence being deleted, the bill run summary was not being updated. We traced this issue to DeleteInvoiceService, which only updates the summary if the bill run status is generated. However, due to our recent changes (#559 #575), calling this service from within DeleteLicenceService means that the bill run's status is pending at the point that it's checked. We therefore update DeleteInvoiceService to accept an optional bill run status (defaulting to the current status) and pass in the bill run's initial status so it can correctly check whether the bill run is generated.