Create migration
We create a migration that adds a new `ruleset` column to the `bill_runs` table. Since there are only presroc bill runs in the table, we want this to default to `presroc` for all existing entries. However, when creating new bill runs, we want the default to be `sroc`. We therefore create the column with the default `presroc` then immediately alter the default to `sroc`.