Add `presroc`/`sroc` requirement to validation
The acceptance criteria call for us to only allow ruleset values of `sroc` and `presroc`. We therefore update our validation accordingly. Since it is likely we will want to use these elsewhere, we put the values into `StaticLookupLib` and refer to them. While doing this, we also update the existing region validation to use the valid regions array in `StaticLookupLib`.