Add SRoC support to Create Bill Run (#587)
* Add SRoC support to Create Bill Run

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-89

Our first endpoint supporting SRoC is Create Bill Run.

In addition to the ACs currently on Jira, we also want to ensure the following is covered:
- Since this will be implemented as a `v3` endpoint we want to mark the existing `v2` endpoint as deprecated;
- We are adding a column to the `bill_runs` db table to hold the bill run `ruleset` and all existing bill runs should have this set to `presroc`.

* Create migration

We create a migration that adds a new `ruleset` column to the `bill_runs` table. Since there are only presroc bill runs in the table, we want this to default to `presroc` for all existing entries. However, when creating new bill runs, we want the default to be `sroc`. We therefore create the column with the default `presroc` then immediately alter the default to `sroc`.

* Add v3 route

We add our v3 Create Bill Run route to `bill_run.routes.js`, keeping it pointing to the original endpoint for now.

* Set up versioned endpoints

We rename the existing `create` endpoint `createV2` and point the v2 route to it. We then clone the endpoint as `createV3` and point the v3 route to it.

* Set v2 default for `ruleset`

When the v2 endpoint is called, we want to pass the ruleset value `presroc` to the service. We therefore do this manually within the controller, and create a test to check this is done.

* Correctly set default ruleset in payload

We realised that we were defining the default incorrectly -- as a separate string instead of part of the payload 🤦‍♂️

* Update translator

We update the bill run translator to accept `ruleset` and default it to `sroc` if not specified

* Fix SonarCloud error

* Add v2 guard service

Our pattern is that where possible, we will attempt to validate v2 requests and reject them if they appear to be v3 requests. We do this by creating `CreateBillRunV2GuardService` and calling it from the v2 controller.

* Test that `ruleset` is written to db

Given that we insert the output of the translator into the db we would expect the ruleset to be written, however we add a unit test to confirm this is done.

* Add `presroc`/`sroc` requirement to validation

The acceptance criteria call for us to only allow ruleset values of `sroc` and `presroc`. We therefore update our validation accordingly. Since it is likely we will want to use these elsewhere, we put the values into `StaticLookupLib` and refer to them. While doing this, we also update the existing region validation to use the valid regions array in `StaticLookupLib`.

* Fix validation commit

We didn't actually add the updated `StaticLookupLib` to the previous commit 🙄 We also left `.only` in one of the tests.

* Update unit test for bill run validation

Since our `CreateBillRunService` tests check that our payload doesn't result in an error, we add the ruleset to these tests.

* Mark v2 route as deprecated

We update the v2 route to include deprecation info

* Remove migration default

We created the `ruleset` migration to set the default value to `sroc`. However now that we've developed the rest of the v3 endpoint we realise that we're setting the default ruleset elsewhere so it isn't needed in the migration -- we can simply set it to `notNullable` instead.

* Update bill run helpers

We updated the bill run migration to make `ruleset` a required field but didn't update the defaults in `BillRunHelper and `NewBillRunHelper`to specify the ruleset -- we fix this here.

* Refactor controller tests

We refactor the existing Create Bill Run controller tests to completely stub out the called services, and create a separate set of tests for the v3 route.

* Remove `.only`

* Update all tests with new helpers

We update the existing tests in the controller to use the new test helpers.

* Revert validation naming convention

On reflection, we preferred the clearer approach of having `_validRegions()` and `_validRulesets()` methods that we refer to in `BillRunTranslator`, instead of directly referring to `StaticLookupLib.regions` and `.rulesets` in the validation schema.

* Rename `createV3` endpoint

As pointed out in review, our convention is that the previous endpoint will have the version in the name and the current endpoint won't. We therefore rename `createV3` to just `create`

* Replace bill run in controller test with dummy

As pointed out in review, our controller testing doesn't require an actual bill run, so we replace it with a simple object containing an id.

* Remove `.only`

* Replace bill run in v2 controller test with dummy

Earlier we replaced the created bill run in the v3 controller test with a dummy bill run. We now replace the bill run in the v2 test with a dummy.